feat(deprecation): add ability to deprecate packages
This introduces the ability to deprecate packages. This gets a little
tricky because we are publishing to multiple package managers which each
handle deprecation differently. For example, with NPM you can run an npm
command, but it's not currently possible
to "deprecate" a package on PyPI, and on Nuget it must be done manually.

Because of this, and because npm seems to currently be the source of
truth for CDK Constructs (e.g. Construct Hub) I am only "deprecating"
packages on npm.

When a user adds a package to the list of `deprecateTypes`, it will

1. Add a disclaimer to the package README that the package is deprecated
2. Remove the build/release workflows for the package
3. Add an npm deprecate workflow for the package

It is also possible to un-deprecate a package on npm by running `npm
deprecate` with an empty message. This update also handles that scenario
if an empty string is provided in `deprecateTypes`

fixes #13