chore: switch to superchain Node 20
Releasing this repository is currently broken.

`@aws-cdk/cloud-assembly-schema` accidentally advertises that it
needs node `18.18` or higher to work; Superchain currently vends node
`18.17`, and it's hard to release Superchain.

Because of the way our package is versioned and released, the version
constraint is not easy to fix.

It's easier to switch to a Node 20 version of Superchain, which is
definitely higher than `18.18`, to unblock the releases.